+++ backend/lib/emailTemp.js
@@ -1,0 +1,41 @@
+const nodemailer = require("nodemailer")
+
+const myEmail = process.env.MyEmail
+const myPassword = process.env.MyPassword
+
+var transporter = nodemailer.createTransport({
+  service: "gmaill",
+  port: 465,
+  host: "smtp.gmail.com",
+  auth: {
+    user: myEmail,
+    pass: myPassword,
+  },
+  secure: true,
+})
+
+const htmlContext = (name, newOTP) => {
+  return `<div style="font-family: Helvetica,Arial,sans-serif;min-width:1000px;overflow:auto;line-height:2">
+        <div style="margin:50px auto;width:70%;padding:20px 0">
+          <div style="border-bottom:1px solid #eee">
+            <a href="#" style="font-size:1.4em;color: #00466a;text-decoration:none;font-weight:600">CartZilla</a>
+          </div>
+          <p style="font-size:1.1em">Hi ${name},</p>
+          <p>Thank you for choosing CartZilla. Use the following OTP to complete your Sign Up procedures.</p>
+          <p>Your OTP is</p>
+          <h2 style="background: #00466a;margin: 0 auto;width: max-content;padding: 0 10px;color: #fff;border-radius: 4px;">${newOTP}</h2>
+          <p style="font-size:0.9em;">Regards,<br />CartZilla</p>
+          <hr style="border:none;border-top:1px solid #eee" />
+          <div style="float:right;padding:8px 0;color:#aaa;font-size:0.8em;line-height:1;font-weight:300">
+            <p>CartZilla</p>
+            <p>Lucknow</p>
+            <p>Uttar Pradesh</p>
+          </div>
+        </div>
+      </div>`
+}
+
+module.exports = {
+  transporter,
+  htmlContext,
+}
